sys log highlighter refactorization
